+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add `attributedTo` property to `Emoji` objects.
 - Support integrity proofs with injected context.
+- Serve outgoing public activities.
 
 ### Changed